Allow iteration through an RB-Tree.
Add functions which return the next element with a greater key
than the current element. Also add a functions which return the
previous, first and last element.
This iteration is really slow and is if you need to iterate often
through a tree, you should link the elements in a linked list.