Rewrite build system.
New build system no longer uses autotools. The build system should run
on every system supporting an SH. It is now based on configuration files
which have to be created for every supported system. Some precreated
configuration files are already shipped with ONS and are autodetected by
the build scripts.
A build script for MS-Windows BATCH is also planned.