Implement new asynchio interface.
Created a new asynchio interface based on a single OBJ structure which
contains the current state of an IO object. Currently no backend is
defined so using this object works but will not do any decent work.

However, defining new backends now will be easy work and this allows to
use any kind of IO object with this frontend.