Modify asynchio options interface.
asynchio options are now set and unset with the ctrl function with the
two new constants ASYN_SET and ASYN_UNSET. The first argument of ctrl
will be a biteset of options flags plus the SET or UNSET flag.
If future options require parameters, the order of the parameters will
be fixed and thus it's not recommended to set more than one options with
parameters with this function, except if you care for the order of your
arguments.